CF-Tracker/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
contourpy==1.3.3
cycler==0.12.1
Django==6.0.3
djangorestframework==3.16.1
fonttools==4.61.1
idna==3.11
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pandas==3.0.2
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2025.3
urllib3==2.6.3
